backend/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "twain-backend"
version = "0.1.0"
description = "Twain AI — medical consultation backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.118",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "pydantic-settings>=2.6",
    "pyjwt[crypto]>=2.10",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt<4.1",
    "httpx>=0.28",
    "anthropic>=0.45",
    "weasyprint>=63",
    "python-multipart>=0.0.12",
    "websockets>=13",
    "jinja2>=3.1",
    "email-validator>=2.2",
]

[tool.setuptools]
packages = ["app", "app.routers"]

[tool.setuptools.package-data]
app = ["templates/*.html"]
